Move strip_html to the front of the MapCompose pipeline.
This causes strip_html to be called for each element in the ingredients list,
rather than calling it at the end after all the ingredients have been joined
with newlines.  The advantage to this approach is that it lets us have more
fine-grained control over sanitizing individual ingredients, since it lets us
examine each one free of html.